Renaming optimizer to tuner, separate search from tuner.
